# SDAP-312: Improved UTM reprojection speed